费用标准添加社区

+++ estate-order/src/main/java/com/estate/sdzy/tariff/service/impl/FBillServiceImpl.java
@@ -142,43 +142,51 @@ public class FBillServiceImpl extends ServiceImpl<FBillMapper, FBill> implements
             billQueryWrapper.eq("cost_rule_id", ruleId).eq("account_period", accountPeriod);
             List<FBill> fBills = billMapper.selectList(billQueryWrapper);
             StringBuilder sb = new StringBuilder("");
-
-            for (FBill res : fBills) {
-                String types = "";
-                // 区分车位房产等
-                if ("room".equals(res.getPropertyType())) {
-                    types = "room";
-                }
-                if ("park".equals(res.getPropertyType())) {
-                    types = "park";
-                }
-                if ("an".equals(res.getPropertyType())) {
-                    types = "an";
-                }
-                if ("rq".equals(res.getPropertyType())) {
-                    types = "rq";
-                }
-                if ("water".equals(res.getPropertyType())) {
-                    types = "water";
-                }
-                int i = res.getPayPrice().compareTo(new BigDecimal(0));
-                if ("否".equals(res.getIsPayment()) && i <= 0) {
-                    // 符合条件的先执行删除，在重新生成。
-                    sb.append(res.getId()).append(",");
-                    if (!StringUtils.isEmpty(res.getPropertyId())) {
-                        try {
-                            billMapper.deleteById(res.getId());
-                            CrontabCostRule.execute(Integer.valueOf(ruleId.toString()), types, res.getPropertyId() + "", res.getAccountPeriod());
-                            return true;
-                        } catch (Exception sqlException) {
-                            throw new OrderException(500,"批量重新生成账单异常");
-                        }
-                    }
-                }
-
-
-            }
-
+            // 如果查询结果是空的，表示还没有生成账单。可以手动生成。
+           if(!fBills.isEmpty()){
+               for (FBill res : fBills) {
+                   String types = "";
+                   // 区分车位房产等
+                   if ("room".equals(res.getPropertyType())) {
+                       types = "room";
+                   }
+                   if ("park".equals(res.getPropertyType())) {
+                       types = "park";
+                   }
+                   if ("an".equals(res.getPropertyType())) {
+                       types = "an";
+                   }
+                   if ("rq".equals(res.getPropertyType())) {
+                       types = "rq";
+                   }
+                   if ("water".equals(res.getPropertyType())) {
+                       types = "water";
+                   }
+                   int i = res.getPayPrice().compareTo(new BigDecimal(0));
+                   if ("否".equals(res.getIsPayment()) && i <= 0) {
+                       // 符合条件的先执行删除，在重新生成。
+                       sb.append(res.getId()).append(",");
+                       if (!StringUtils.isEmpty(res.getPropertyId())) {
+                           try {
+                               billMapper.deleteById(res.getId());
+                               CrontabCostRule.execute(Integer.valueOf(ruleId.toString()), types, res.getPropertyId() + "", res.getAccountPeriod());
+                               return true;
+                           } catch (Exception sqlException) {
+                               throw new OrderException(500,"批量重新生成账单异常");
+                           }
+                       }
+                   }
+               }
+           }else{
+               try {
+                   CrontabCostRule.execute(Integer.valueOf(ruleId.toString()), null, null,accountPeriod);
+                   return true;
+               } catch (SQLException sqlException) {
+                   sqlException.printStackTrace();
+               } catch (ClassNotFoundException e) {
+                   e.printStackTrace();
+               }
+           }
         }
         return false;
     }
